--- src/interface_adapter/NextDepartures/NextDeparturesState.java
@@ -1,10 +1,0 @@
-package interface_adapter.NextDepartures;
-
-import java.util.HashMap;
-
-public class NextDeparturesState {
-
-    private HashMap nextDepartures = null;
-
-
-}
