+++ src/interface_adapter/next_departures/NextDeparturesState.java
@@ -1,0 +1,5 @@
+package interface_adapter.next_departures;
+
+public class NextDeparturesState {
+
+}
