+++ src/interface_adapter/next_departures/NextDeparturesViewModel.java
@@ -1,16 +1,11 @@
-package interface_adapter.NextDepartures;
+package interface_adapter.next_departures;
 
 import use_case.next_departures.NextDepartureOutputData;
 
 public class NextDeparturesViewModel {
-
     public static void getNextDepartures(NextDepartureOutputData station) {
     }
 
     public static void getDeparturesError(NextDepartureOutputData error) {
-    }
-
-    public static NextDeparturesState getState() {
-        return nextDepartures;
     }
 }
